Extract evolutionary rates from a Bayesian clock tree

Extract evolutionary rate summary statistics for each node from a Bayesian clock summary tree produced by Mr. Bayes and stores them in a data frame.

Usage

get_clockrate_table(tree, summary = "median",
                    drop_dummyextant = TRUE)

Arguments

tree

An S4 class object of type treedata; a Bayesian clock tree imported using treeio::read.mrbayes for Mr. Bayes summary trees.

summary

The name of the rate summary. Should be one of "mean" or "median".

drop_dummyextant

logical; whether to drop a dummy extant tip labeled as "Dummyextant" from the Mr. Bayes summary tree prior to extracting the clock rates (when present). Default is TRUE.

Value

A data frame with a column containing the node identifer (node) and one column for each relaxed clock partition in the tree object containing clock rates.

See also

Examples

# See vignette("rates-selection") for how to use this
# function as part of an analysis pipeline

data("tree3p")

rate_table <- get_clockrate_table(tree3p)

head(rate_table)
#>   nodes   rates1   rates2   rates3
#> 1     1 0.927276 0.958621 0.984651
#> 2     2 1.039289 0.757550 0.899578
#> 3     3 1.151933 0.627024 0.795243
#> 4     4 1.189229 0.451265 0.693551
#> 5     5 1.188293 0.446332 0.693501
#> 6     6 0.649204 0.694196 0.656767