Posterior parameter samples (single clock)

An example dataset of posterior parameter samples resulting from a clock-based Bayesian inference analysis using the skyline fossilized birth–death process (FBD) tree model with Mr. Bayes after combining all parameter (.p) files into a single data frame with the combine_log function. This particular example was produced by analyzing the data set with a single morphological partition from Simões & Pierce (2021).

Usage

data("posterior1p")

Format

A data frame with 4000 observations on several variables estimated for each generation during analysis:

Gen

A numeric vector for the generation number

LnL

A numeric vector for the natural log likelihood of the cold chain

LnPr

A numeric vector for the natural log likelihood of the priors

TH

A numeric vector for the total tree height (sum of all branch durations, as chronological units)

TL

A numeric vector for total tree length (sum of all branch lengths, as accumulated substitutions/changes)

prop_ancfossil

A numeric vector indicating the proportion of fossils recovered as ancestors

sigma

A numeric vector for the standard deviation of the lognormal distribution governing how much rates vary across characters.

net_speciation_1, net_speciation_2, net_speciation_3, net_speciation_4

A numeric vector for net speciation estimates for each time bin

relative_extinction_1, relative_extinction_2, relative_extinction_3, relative_extinction_4

A numeric vector for relative extinction estimates for each time bin

relative_fossilization_1, relative_fossilization_2, relative_fossilization_3, relative_fossilization_4

A numeric vector for relative fossilization estimates for each time bin

tk02var

A numeric vector for the variance on the base of the clock rate

clockrate

A numeric vector for the base of the clock rate

Details

Datasets like this one can be produced from parameter log (.p) files using combine_log. The number of variables depends on parameter set up, but for clock analyses with Mr. Bayes, will typically include the ones above, possibly also including alpha, which contains the shape of the gamma distribution governing how much rates vary across characters. When using the traditional FBD model rather than the skyline FBD model used to produce this dataset, there will be only one column for each of net_speciation, relative_extinction and relative_fossilization. When using more than one morphological partition, different columns may be present; see posterior3p for an example with 3 partitions.

See also

References

Simões, T. R. and S. E. Pierce (2021). Sustained High Rates of Morphological Evolution During the Rise of Tetrapods. Nature Ecology & Evolution 5: 1403–1414.