Mean clock rates by node and clade (3 clock partitions)

A data set containing the mean clock rates for a tree with 3 clock partitions, such as the output of get_clockrate_table_MrBayes but with an additional "clade" column added, which is required for use in clockrate_summary and clockrate_dens_plot.

Usage

data("rate_table_clades_means3")

Format

A data frame with 79 observations on the following 5 variables.

clade

A character vector containing the clade names for each corresponding node

nodes

A numeric vector for the node numbers in the summary tree

rates1

A numeric vector containing the mean posterior clock rate for each node for the first partition

rates2

A numeric vector containing the mean posterior clock rate for each node for the second partition

rates3

A numeric vector containing the mean posterior clock rate for each node for the third partition

Details

rate_table_clades_means3 was created by running get_clockrate_table_MrBayes(tree3p) and then adding a "clade" column. It can be produced by using the following procedure:

1) Import tree file:

data("tree3p")

2) Produce clock rate table with, for instance, mean rate values from each branch in the tree:

rate_table <- get_clockrate_table_MrBayes(tree3p, summary = "mean")

write.csv(rate_table, file = "rate_table.csv", row.names = FALSE)

3) Now, manually add clades using, e.g., Excel:

3.1) Manually edit rate_table.csv, adding a "clade" column. This introduces customized clade names to individual nodes in the tree.

3.2) Save the edited rate table with a different name to differentiate from the original output (e.g., rate_table_clades_means.csv).

4) Read the file back in:

rate_table_clades_means3 <- read.csv("rate_table_clades_means.csv")

head(rate_table_clades_means3)

See also