Conduct pairwise t-tests between node rates and clock base rate

Produces a data frame containing the results of 1-sample t-tests for the mean of posterior clock rates against each node's absolute clock rate.

Usage

get_pwt_rates(rate_table, posterior)

Arguments

rate_table

A data frame containing a single "value" column (for all rate values) and one column for the "clock" variable (indicating to which clock partition each rate values refers to), such as from the output of get_clockrate_table with an extra clade column added, and followed by clock_reshape.

posterior

A data frame of posterior parameter estimates including a "clockrate" column indicating the base of the clock rate estimate for each generation that will be used for pairwise t-tests. Such data frame can be imported using combine_log (no need to reshape from wide to long). See the posterior1p or posterior3p datasets for an examples of how the input file should look.

Details

get_pwt_rates() first transforms relative clock rates to absolute rate values for each node and each clock, by multiplying these by the mean posterior clock rate base value. Then, for each node and clock, a one-sample t-test is performed with the null hypothesis that the mean of the posterior clockrates is equal to that node and clock's absolute clock rate.

Value

A long data frame with one row per node per clock and the following columns:

clade

The name of the clade, taken from the "clade" column of rate_table

nodes

The node number, taken from the "node" column of rate_table

clock

The clock partition number

relative rate

The relative mean clock rate per node, taken from the "rates" columns of rate_table

absolute rate (mean)

The absolute mean clock rate per node; the relative clock rate multiplied by the mean of the posterior clock rates

null

The absolute clock rate used as the null value in the t-test

p.value

The p-value of the test comparing the mean ofthe posterior clockrates to each absolute clockrate

See also

Examples

# See vignette("rates-selection") for how to use this
# function as part of an analysis pipeline

# Load example rate table and posterior data sets
data("rate_table_clades_means3")
data("posterior3p")

get_pwt_rates(rate_table_clades_means3, posterior3p)
#>                 clade nodes clock relative rate absolute rate (mean)
#> 1         Dipnomorpha     1     1      0.943696         0.0118443306
#> 2         Dipnomorpha     2     1      1.065326         0.0133709091
#> 3         Dipnomorpha     3     1      1.182460         0.0148410582
#> 4         Dipnomorpha     4     1      1.229767         0.0154348084
#> 5         Dipnomorpha     5     1      1.230564         0.0154448116
#> 6               Other     6     1      0.658855         0.0082692906
#> 7               Other     7     1      0.603090         0.0075693840
#> 8      Osteolepididae     8     1      0.843373         0.0105851764
#> 9      Osteolepididae     9     1      0.872012         0.0109446246
#> 10     Osteolepididae    10     1      0.811473         0.0101847995
#> 11     Osteolepididae    11     1      0.844565         0.0106001372
#> 12     Osteolepididae    12     1      0.827772         0.0103893683
#> 13     Osteolepididae    13     1      0.812783         0.0102012413
#> 14      Rhizodontidae    14     1      0.478512         0.0060058052
#> 15      Rhizodontidae    15     1      0.280376         0.0035189998
#> 16      Rhizodontidae    16     1      0.222560         0.0027933511
#> 17      Rhizodontidae    17     1      0.119009         0.0014936822
#> 18      Rhizodontidae    18     1      0.030353         0.0003809606
#> 19      Rhizodontidae    19     1      0.499087         0.0062640421
#> 20      Rhizodontidae    20     1      0.551223         0.0069184011
#> 21  Tristichopteridae    21     1      0.650066         0.0081589798
#> 22  Tristichopteridae    22     1      0.614651         0.0077144861
#> 23  Tristichopteridae    23     1      0.681184         0.0085495419
#> 24  Tristichopteridae    24     1      0.710585         0.0089185540
#> 25  Tristichopteridae    25     1      0.722366         0.0090664173
#> 26  Tristichopteridae    26     1      0.727056         0.0091252815
#> 27  Tristichopteridae    27     1      0.741828         0.0093106849
#> 28  Tristichopteridae    28     1      0.736285         0.0092411147
#> 29  Tristichopteridae    29     1      0.694766         0.0087200097
#> 30      Rhizodontidae    30     1      0.889881         0.0111688985
#> 31  Tristichopteridae    31     1      0.781584         0.0098096626
#> 32  Tristichopteridae    32     1      0.723329         0.0090785039
#> 33     Elpisostegalia    33     1      1.423592         0.0178675064
#> 34     Elpisostegalia    34     1      1.410658         0.0177051718
#> 35     Elpisostegalia    35     1      1.450903         0.0182102869
#> 36     Elpisostegalia    36     1      1.754826         0.0220248251
#> 37     Elpisostegalia    37     1      1.794746         0.0225258611
#> 38     Elpisostegalia    38     1      1.806100         0.0226683652
#> 39     Elpisostegalia    39     1      1.806062         0.0226678883
#> 40              Other    40     1      1.806719         0.0226761343
#> 41              Other    41     1      1.000000         0.0125510023
#> 42        Dipnomorpha    42     1      0.990027         0.0124258311
#> 43        Dipnomorpha    43     1      1.029762         0.0129245452
#> 44        Dipnomorpha    44     1      1.131153         0.0141971039
#> 45              Other    45     1      1.206840         0.0151470516
#> 46              Other    46     1      0.881218         0.0110601691
#> 47              Other    47     1      0.718667         0.0090199912
#> 48              Other    48     1      0.782346         0.0098192264
#> 49              Other    49     1      0.747936         0.0093873464
#> 50      Rhizodontidae    50     1      0.783704         0.0098362707
#> 51      Rhizodontidae    51     1      0.833884         0.0104660800
#> 52      Rhizodontidae    52     1      0.842345         0.0105722740
#> 53      Rhizodontidae    53     1      0.838206         0.0105203254
#> 54      Rhizodontidae    54     1      0.834469         0.0104734223
#> 55      Rhizodontidae    55     1      0.803287         0.0100820570
#> 56     Osteolepididae    56     1      0.644209         0.0080854686
#> 57     Osteolepididae    57     1      0.509998         0.0064009861
#> 58     Osteolepididae    58     1      0.374429         0.0046994592
#> 59     Osteolepididae    59     1      0.251536         0.0031570289
#> 60     Osteolepididae    60     1      0.164795         0.0020683424
#> 61              Other    61     1      0.540788         0.0067874314
#> 62              Other    62     1      0.793012         0.0099530954
#> 63  Tristichopteridae    63     1      0.749319         0.0094047045
#> 64  Tristichopteridae    64     1      0.683167         0.0085744306
#> 65  Tristichopteridae    65     1      0.691132         0.0086743993
#> 66  Tristichopteridae    66     1      0.712150         0.0089381963
#> 67  Tristichopteridae    67     1      0.720689         0.0090453693
#> 68  Tristichopteridae    68     1      0.721891         0.0090604556
#> 69  Tristichopteridae    69     1      0.739068         0.0092760442
#> 70  Tristichopteridae    70     1      0.739042         0.0092757178
#> 71  Tristichopteridae    71     1      0.742390         0.0093177386
#> 72  Tristichopteridae    72     1      0.715251         0.0089771169
#> 73     Elpisostegalia    73     1      1.132515         0.0142141983
#> 74     Elpisostegalia    74     1      1.452664         0.0182323892
#> 75     Elpisostegalia    75     1      1.460614         0.0183321696
#> 76     Elpisostegalia    76     1      1.609960         0.0202066116
#> 77     Elpisostegalia    77     1      1.774477         0.0222714649
#> 78     Elpisostegalia    78     1      1.799551         0.0225861687
#> 79     Elpisostegalia    79     1      1.797607         0.0225617696
#> 80        Dipnomorpha     1     2      0.981486         0.0123186330
#> 81        Dipnomorpha     2     2      0.772074         0.0096903025
#> 82        Dipnomorpha     3     2      0.656872         0.0082444020
#> 83        Dipnomorpha     4     2      0.523709         0.0065730729
#> 84        Dipnomorpha     5     2      0.517773         0.0064985701
#> 85              Other     6     2      0.717277         0.0090025453
#> 86              Other     7     2      0.718170         0.0090137533
#> 87     Osteolepididae     8     2      0.461855         0.0057967432
#> 88     Osteolepididae     9     2      0.531676         0.0066730667
#> 89     Osteolepididae    10     2      0.381710         0.0047908431
#> 90     Osteolepididae    11     2      0.348332         0.0043719157
#> 91     Osteolepididae    12     2      0.419603         0.0052664382
#> 92     Osteolepididae    13     2      0.441795         0.0055449701
#> 93      Rhizodontidae    14     2      0.334641         0.0042000800
#> 94      Rhizodontidae    15     2      0.120787         0.0015159979
#> 95      Rhizodontidae    16     2      0.076851         0.0009645571
#> 96      Rhizodontidae    17     2      0.050740         0.0006368379
#> 97      Rhizodontidae    18     2      0.019425         0.0002438032
#> 98      Rhizodontidae    19     2      0.357528         0.0044873347
#> 99      Rhizodontidae    20     2      0.390546         0.0049017437
#> 100 Tristichopteridae    21     2      0.422987         0.0053089108
#> 101 Tristichopteridae    22     2      0.339551         0.0042617054
#> 102 Tristichopteridae    23     2      0.312348         0.0039202805
#> 103 Tristichopteridae    24     2      0.269648         0.0033843527
#> 104 Tristichopteridae    25     2      0.298689         0.0037488463
#> 105 Tristichopteridae    26     2      0.324625         0.0040743691
#> 106 Tristichopteridae    27     2      0.328442         0.0041222763
#> 107 Tristichopteridae    28     2      0.332024         0.0041672340
#> 108 Tristichopteridae    29     2      0.315428         0.0039589375
#> 109     Rhizodontidae    30     2      0.556978         0.0069906321
#> 110 Tristichopteridae    31     2      0.326863         0.0041024583
#> 111 Tristichopteridae    32     2      0.328587         0.0041240962
#> 112    Elpisostegalia    33     2      1.405726         0.0176432702
#> 113    Elpisostegalia    34     2      1.415251         0.0177628185
#> 114    Elpisostegalia    35     2      1.414526         0.0177537191
#> 115    Elpisostegalia    36     2      1.384021         0.0173708507
#> 116    Elpisostegalia    37     2      1.352855         0.0169796862
#> 117    Elpisostegalia    38     2      1.350512         0.0169502792
#> 118    Elpisostegalia    39     2      1.360204         0.0170719235
#> 119             Other    40     2      1.351978         0.0169686790
#> 120             Other    41     2      1.000000         0.0125510023
#> 121       Dipnomorpha    42     2      0.970792         0.0121844126
#> 122       Dipnomorpha    43     2      0.864016         0.0108442668
#> 123       Dipnomorpha    44     2      0.721901         0.0090605811
#> 124             Other    45     2      0.591140         0.0074193995
#> 125             Other    46     2      0.837369         0.0105098202
#> 126             Other    47     2      0.726465         0.0091178639
#> 127             Other    48     2      0.678900         0.0085208754
#> 128             Other    49     2      0.581344         0.0072964499
#> 129     Rhizodontidae    50     2      0.523987         0.0065765620
#> 130     Rhizodontidae    51     2      0.508899         0.0063871925
#> 131     Rhizodontidae    52     2      0.450616         0.0056556824
#> 132     Rhizodontidae    53     2      0.421772         0.0052936613
#> 133     Rhizodontidae    54     2      0.377759         0.0047412541
#> 134     Rhizodontidae    55     2      0.391706         0.0049163029
#> 135    Osteolepididae    56     2      0.473161         0.0059386448
#> 136    Osteolepididae    57     2      0.348636         0.0043757312
#> 137    Osteolepididae    58     2      0.207720         0.0026070942
#> 138    Osteolepididae    59     2      0.109647         0.0013761797
#> 139    Osteolepididae    60     2      0.065865         0.0008266718
#> 140             Other    61     2      0.388379         0.0048745457
#> 141             Other    62     2      0.628516         0.0078885058
#> 142 Tristichopteridae    63     2      0.545083         0.0068413380
#> 143 Tristichopteridae    64     2      0.424648         0.0053297580
#> 144 Tristichopteridae    65     2      0.370732         0.0046530582
#> 145 Tristichopteridae    66     2      0.329625         0.0041371241
#> 146 Tristichopteridae    67     2      0.306807         0.0038507354
#> 147 Tristichopteridae    68     2      0.306247         0.0038437068
#> 148 Tristichopteridae    69     2      0.322208         0.0040440333
#> 149 Tristichopteridae    70     2      0.325560         0.0040861043
#> 150 Tristichopteridae    71     2      0.331972         0.0041665813
#> 151 Tristichopteridae    72     2      0.327293         0.0041078552
#> 152    Elpisostegalia    73     2      1.028607         0.0129100488
#> 153    Elpisostegalia    74     2      1.423276         0.0178635403
#> 154    Elpisostegalia    75     2      1.418243         0.0178003711
#> 155    Elpisostegalia    76     2      1.401199         0.0175864518
#> 156    Elpisostegalia    77     2      1.370817         0.0172051273
#> 157    Elpisostegalia    78     2      1.361404         0.0170869847
#> 158    Elpisostegalia    79     2      1.360754         0.0170788266
#> 159       Dipnomorpha     1     3      1.006164         0.0126283667
#> 160       Dipnomorpha     2     3      0.913194         0.0114615000
#> 161       Dipnomorpha     3     3      0.813618         0.0102117214
#> 162       Dipnomorpha     4     3      0.722519         0.0090683376
#> 163       Dipnomorpha     5     3      0.720479         0.0090427336
#> 164             Other     6     3      0.663950         0.0083332380
#> 165             Other     7     3      0.629632         0.0079025127
#> 166    Osteolepididae     8     3      0.419218         0.0052616061
#> 167    Osteolepididae     9     3      0.412089         0.0051721300
#> 168    Osteolepididae    10     3      0.438156         0.0054992970
#> 169    Osteolepididae    11     3      0.332682         0.0041754925
#> 170    Osteolepididae    12     3      0.305413         0.0038332393
#> 171    Osteolepididae    13     3      0.308068         0.0038665622
#> 172     Rhizodontidae    14     3      0.308952         0.0038776573
#> 173     Rhizodontidae    15     3      0.170589         0.0021410629
#> 174     Rhizodontidae    16     3      0.082334         0.0010333742
#> 175     Rhizodontidae    17     3      0.052304         0.0006564676
#> 176     Rhizodontidae    18     3      0.021829         0.0002739758
#> 177     Rhizodontidae    19     3      0.435641         0.0054677312
#> 178     Rhizodontidae    20     3      0.520903         0.0065378547
#> 179 Tristichopteridae    21     3      0.553043         0.0069412440
#> 180 Tristichopteridae    22     3      0.528972         0.0066391288
#> 181 Tristichopteridae    23     3      0.597280         0.0074964626
#> 182 Tristichopteridae    24     3      0.515361         0.0064682971
#> 183 Tristichopteridae    25     3      0.491891         0.0061737251
#> 184 Tristichopteridae    26     3      0.399800         0.0050178907
#> 185 Tristichopteridae    27     3      0.435036         0.0054601378
#> 186 Tristichopteridae    28     3      0.400769         0.0050300526
#> 187 Tristichopteridae    29     3      0.374070         0.0046949534
#> 188     Rhizodontidae    30     3      0.404906         0.0050819761
#> 189 Tristichopteridae    31     3      0.638693         0.0080162373
#> 190 Tristichopteridae    32     3      0.620576         0.0077888508
#> 191    Elpisostegalia    33     3      0.636750         0.0079918507
#> 192    Elpisostegalia    34     3      0.626356         0.0078613956
#> 193    Elpisostegalia    35     3      0.674748         0.0084687637
#> 194    Elpisostegalia    36     3      0.958504         0.0120301859
#> 195    Elpisostegalia    37     3      0.981413         0.0123177168
#> 196    Elpisostegalia    38     3      0.996104         0.0125021036
#> 197    Elpisostegalia    39     3      1.001471         0.0125694648
#> 198             Other    40     3      1.001633         0.0125714981
#> 199             Other    41     3      1.000000         0.0125510023
#> 200       Dipnomorpha    42     3      0.982593         0.0123325270
#> 201       Dipnomorpha    43     3      0.936269         0.0117511144
#> 202       Dipnomorpha    44     3      0.860974         0.0108060866
#> 203             Other    45     3      0.769013         0.0096518839
#> 204             Other    46     3      0.857469         0.0107620954
#> 205             Other    47     3      0.706594         0.0088684629
#> 206             Other    48     3      0.698728         0.0087697367
#> 207             Other    49     3      0.591015         0.0074178306
#> 208     Rhizodontidae    50     3      0.478347         0.0060037343
#> 209     Rhizodontidae    51     3      0.423559         0.0053160900
#> 210     Rhizodontidae    52     3      0.421364         0.0052885405
#> 211     Rhizodontidae    53     3      0.431023         0.0054097707
#> 212     Rhizodontidae    54     3      0.385438         0.0048376332
#> 213     Rhizodontidae    55     3      0.320949         0.0040282316
#> 214    Osteolepididae    56     3      0.486264         0.0061031006
#> 215    Osteolepididae    57     3      0.363379         0.0045607707
#> 216    Osteolepididae    58     3      0.243299         0.0030536463
#> 217    Osteolepididae    59     3      0.135198         0.0016968704
#> 218    Osteolepididae    60     3      0.074238         0.0009317613
#> 219             Other    61     3      0.466393         0.0058536996
#> 220             Other    62     3      0.651189         0.0081730746
#> 221 Tristichopteridae    63     3      0.619467         0.0077749317
#> 222 Tristichopteridae    64     3      0.583625         0.0073250787
#> 223 Tristichopteridae    65     3      0.582727         0.0073138079
#> 224 Tristichopteridae    66     3      0.591158         0.0074196254
#> 225 Tristichopteridae    67     3      0.586489         0.0073610248
#> 226 Tristichopteridae    68     3      0.563005         0.0070662770
#> 227 Tristichopteridae    69     3      0.536443         0.0067328973
#> 228 Tristichopteridae    70     3      0.472132         0.0059257298
#> 229 Tristichopteridae    71     3      0.424319         0.0053256287
#> 230 Tristichopteridae    72     3      0.405045         0.0050837207
#> 231    Elpisostegalia    73     3      0.670266         0.0084125101
#> 232    Elpisostegalia    74     3      0.662369         0.0083133948
#> 233    Elpisostegalia    75     3      0.689177         0.0086498621
#> 234    Elpisostegalia    76     3      0.825191         0.0103569741
#> 235    Elpisostegalia    77     3      0.971535         0.0121937380
#> 236    Elpisostegalia    78     3      0.989252         0.0124161041
#> 237    Elpisostegalia    79     3      0.994892         0.0124868918
#>             null       p.value
#> 1   0.0118443306 3.717002e-108
#> 2   0.0133709091 9.850162e-144
#> 3   0.0148410582  0.000000e+00
#> 4   0.0154348084  0.000000e+00
#> 5   0.0154448116  0.000000e+00
#> 6   0.0082692906  0.000000e+00
#> 7   0.0075693840  0.000000e+00
#> 8   0.0105851764  0.000000e+00
#> 9   0.0109446246  0.000000e+00
#> 10  0.0101847995  0.000000e+00
#> 11  0.0106001372  0.000000e+00
#> 12  0.0103893683  0.000000e+00
#> 13  0.0102012413  0.000000e+00
#> 14  0.0060058052  0.000000e+00
#> 15  0.0035189998  0.000000e+00
#> 16  0.0027933511  0.000000e+00
#> 17  0.0014936822  0.000000e+00
#> 18  0.0003809606  0.000000e+00
#> 19  0.0062640421  0.000000e+00
#> 20  0.0069184011  0.000000e+00
#> 21  0.0081589798  0.000000e+00
#> 22  0.0077144861  0.000000e+00
#> 23  0.0085495419  0.000000e+00
#> 24  0.0089185540  0.000000e+00
#> 25  0.0090664173  0.000000e+00
#> 26  0.0091252815  0.000000e+00
#> 27  0.0093106849  0.000000e+00
#> 28  0.0092411147  0.000000e+00
#> 29  0.0087200097  0.000000e+00
#> 30  0.0111688985  0.000000e+00
#> 31  0.0098096626  0.000000e+00
#> 32  0.0090785039  0.000000e+00
#> 33  0.0178675064  0.000000e+00
#> 34  0.0177051718  0.000000e+00
#> 35  0.0182102869  0.000000e+00
#> 36  0.0220248251  0.000000e+00
#> 37  0.0225258611  0.000000e+00
#> 38  0.0226683652  0.000000e+00
#> 39  0.0226678883  0.000000e+00
#> 40  0.0226761343  0.000000e+00
#> 41  0.0125510023  1.000000e+00
#> 42  0.0124258311  7.497531e-05
#> 43  0.0129245452  4.888080e-32
#> 44  0.0141971039  0.000000e+00
#> 45  0.0151470516  0.000000e+00
#> 46  0.0110601691  0.000000e+00
#> 47  0.0090199912  0.000000e+00
#> 48  0.0098192264  0.000000e+00
#> 49  0.0093873464  0.000000e+00
#> 50  0.0098362707  0.000000e+00
#> 51  0.0104660800  0.000000e+00
#> 52  0.0105722740  0.000000e+00
#> 53  0.0105203254  0.000000e+00
#> 54  0.0104734223  0.000000e+00
#> 55  0.0100820570  0.000000e+00
#> 56  0.0080854686  0.000000e+00
#> 57  0.0064009861  0.000000e+00
#> 58  0.0046994592  0.000000e+00
#> 59  0.0031570289  0.000000e+00
#> 60  0.0020683424  0.000000e+00
#> 61  0.0067874314  0.000000e+00
#> 62  0.0099530954  0.000000e+00
#> 63  0.0094047045  0.000000e+00
#> 64  0.0085744306  0.000000e+00
#> 65  0.0086743993  0.000000e+00
#> 66  0.0089381963  0.000000e+00
#> 67  0.0090453693  0.000000e+00
#> 68  0.0090604556  0.000000e+00
#> 69  0.0092760442  0.000000e+00
#> 70  0.0092757178  0.000000e+00
#> 71  0.0093177386  0.000000e+00
#> 72  0.0089771169  0.000000e+00
#> 73  0.0142141983  0.000000e+00
#> 74  0.0182323892  0.000000e+00
#> 75  0.0183321696  0.000000e+00
#> 76  0.0202066116  0.000000e+00
#> 77  0.0222714649  0.000000e+00
#> 78  0.0225861687  0.000000e+00
#> 79  0.0225617696  0.000000e+00
#> 80  0.0123186330  2.069792e-13
#> 81  0.0096903025  0.000000e+00
#> 82  0.0082444020  0.000000e+00
#> 83  0.0065730729  0.000000e+00
#> 84  0.0064985701  0.000000e+00
#> 85  0.0090025453  0.000000e+00
#> 86  0.0090137533  0.000000e+00
#> 87  0.0057967432  0.000000e+00
#> 88  0.0066730667  0.000000e+00
#> 89  0.0047908431  0.000000e+00
#> 90  0.0043719157  0.000000e+00
#> 91  0.0052664382  0.000000e+00
#> 92  0.0055449701  0.000000e+00
#> 93  0.0042000800  0.000000e+00
#> 94  0.0015159979  0.000000e+00
#> 95  0.0009645571  0.000000e+00
#> 96  0.0006368379  0.000000e+00
#> 97  0.0002438032  0.000000e+00
#> 98  0.0044873347  0.000000e+00
#> 99  0.0049017437  0.000000e+00
#> 100 0.0053089108  0.000000e+00
#> 101 0.0042617054  0.000000e+00
#> 102 0.0039202805  0.000000e+00
#> 103 0.0033843527  0.000000e+00
#> 104 0.0037488463  0.000000e+00
#> 105 0.0040743691  0.000000e+00
#> 106 0.0041222763  0.000000e+00
#> 107 0.0041672340  0.000000e+00
#> 108 0.0039589375  0.000000e+00
#> 109 0.0069906321  0.000000e+00
#> 110 0.0041024583  0.000000e+00
#> 111 0.0041240962  0.000000e+00
#> 112 0.0176432702  0.000000e+00
#> 113 0.0177628185  0.000000e+00
#> 114 0.0177537191  0.000000e+00
#> 115 0.0173708507  0.000000e+00
#> 116 0.0169796862  0.000000e+00
#> 117 0.0169502792  0.000000e+00
#> 118 0.0170719235  0.000000e+00
#> 119 0.0169686790  0.000000e+00
#> 120 0.0125510023  1.000000e+00
#> 121 0.0121844126  6.328562e-31
#> 122 0.0108442668  0.000000e+00
#> 123 0.0090605811  0.000000e+00
#> 124 0.0074193995  0.000000e+00
#> 125 0.0105098202  0.000000e+00
#> 126 0.0091178639  0.000000e+00
#> 127 0.0085208754  0.000000e+00
#> 128 0.0072964499  0.000000e+00
#> 129 0.0065765620  0.000000e+00
#> 130 0.0063871925  0.000000e+00
#> 131 0.0056556824  0.000000e+00
#> 132 0.0052936613  0.000000e+00
#> 133 0.0047412541  0.000000e+00
#> 134 0.0049163029  0.000000e+00
#> 135 0.0059386448  0.000000e+00
#> 136 0.0043757312  0.000000e+00
#> 137 0.0026070942  0.000000e+00
#> 138 0.0013761797  0.000000e+00
#> 139 0.0008266718  0.000000e+00
#> 140 0.0048745457  0.000000e+00
#> 141 0.0078885058  0.000000e+00
#> 142 0.0068413380  0.000000e+00
#> 143 0.0053297580  0.000000e+00
#> 144 0.0046530582  0.000000e+00
#> 145 0.0041371241  0.000000e+00
#> 146 0.0038507354  0.000000e+00
#> 147 0.0038437068  0.000000e+00
#> 148 0.0040440333  0.000000e+00
#> 149 0.0040861043  0.000000e+00
#> 150 0.0041665813  0.000000e+00
#> 151 0.0041078552  0.000000e+00
#> 152 0.0129100488  9.663870e-30
#> 153 0.0178635403  0.000000e+00
#> 154 0.0178003711  0.000000e+00
#> 155 0.0175864518  0.000000e+00
#> 156 0.0172051273  0.000000e+00
#> 157 0.0170869847  0.000000e+00
#> 158 0.0170788266  0.000000e+00
#> 159 0.0126283667  1.436137e-02
#> 160 0.0114615000 2.709439e-246
#> 161 0.0102117214  0.000000e+00
#> 162 0.0090683376  0.000000e+00
#> 163 0.0090427336  0.000000e+00
#> 164 0.0083332380  0.000000e+00
#> 165 0.0079025127  0.000000e+00
#> 166 0.0052616061  0.000000e+00
#> 167 0.0051721300  0.000000e+00
#> 168 0.0054992970  0.000000e+00
#> 169 0.0041754925  0.000000e+00
#> 170 0.0038332393  0.000000e+00
#> 171 0.0038665622  0.000000e+00
#> 172 0.0038776573  0.000000e+00
#> 173 0.0021410629  0.000000e+00
#> 174 0.0010333742  0.000000e+00
#> 175 0.0006564676  0.000000e+00
#> 176 0.0002739758  0.000000e+00
#> 177 0.0054677312  0.000000e+00
#> 178 0.0065378547  0.000000e+00
#> 179 0.0069412440  0.000000e+00
#> 180 0.0066391288  0.000000e+00
#> 181 0.0074964626  0.000000e+00
#> 182 0.0064682971  0.000000e+00
#> 183 0.0061737251  0.000000e+00
#> 184 0.0050178907  0.000000e+00
#> 185 0.0054601378  0.000000e+00
#> 186 0.0050300526  0.000000e+00
#> 187 0.0046949534  0.000000e+00
#> 188 0.0050819761  0.000000e+00
#> 189 0.0080162373  0.000000e+00
#> 190 0.0077888508  0.000000e+00
#> 191 0.0079918507  0.000000e+00
#> 192 0.0078613956  0.000000e+00
#> 193 0.0084687637  0.000000e+00
#> 194 0.0120301859  2.997544e-60
#> 195 0.0123177168  1.667195e-13
#> 196 0.0125021036  1.217487e-01
#> 197 0.0125694648  5.590150e-01
#> 198 0.0125714981  5.165628e-01
#> 199 0.0125510023  1.000000e+00
#> 200 0.0123325270  4.979532e-12
#> 201 0.0117511144 4.184481e-137
#> 202 0.0108060866  0.000000e+00
#> 203 0.0096518839  0.000000e+00
#> 204 0.0107620954  0.000000e+00
#> 205 0.0088684629  0.000000e+00
#> 206 0.0087697367  0.000000e+00
#> 207 0.0074178306  0.000000e+00
#> 208 0.0060037343  0.000000e+00
#> 209 0.0053160900  0.000000e+00
#> 210 0.0052885405  0.000000e+00
#> 211 0.0054097707  0.000000e+00
#> 212 0.0048376332  0.000000e+00
#> 213 0.0040282316  0.000000e+00
#> 214 0.0061031006  0.000000e+00
#> 215 0.0045607707  0.000000e+00
#> 216 0.0030536463  0.000000e+00
#> 217 0.0016968704  0.000000e+00
#> 218 0.0009317613  0.000000e+00
#> 219 0.0058536996  0.000000e+00
#> 220 0.0081730746  0.000000e+00
#> 221 0.0077749317  0.000000e+00
#> 222 0.0073250787  0.000000e+00
#> 223 0.0073138079  0.000000e+00
#> 224 0.0074196254  0.000000e+00
#> 225 0.0073610248  0.000000e+00
#> 226 0.0070662770  0.000000e+00
#> 227 0.0067328973  0.000000e+00
#> 228 0.0059257298  0.000000e+00
#> 229 0.0053256287  0.000000e+00
#> 230 0.0050837207  0.000000e+00
#> 231 0.0084125101  0.000000e+00
#> 232 0.0083133948  0.000000e+00
#> 233 0.0086498621  0.000000e+00
#> 234 0.0103569741  0.000000e+00
#> 235 0.0121937380  1.825627e-29
#> 236 0.0124161041  1.977881e-05
#> 237 0.0124868918  4.247976e-02